validate ID

+++ infra/db/task/task.go
@@ -1,6 +1,7 @@
 package task
 
 import (
+	"fmt"
 	"time"
 
 	taskModel "github.com/team-gleam/kiwi-basket/domain/model/task"
@@ -17,6 +18,11 @@ func NewTaskRepository(h *handler.DbHandler) taskRepository.ITaskRepository {
 	return &TaskRepository{h}
 }
 
+const (
+	IDIsNotZero = "ID is not zero"
+	InvalidID   = "Invalid ID"
+)
+
 type taskDB struct {
 	ID       uint `gorm:"primary_key;auto_increment"`
 	Username string
@@ -25,6 +31,10 @@ type taskDB struct {
 }
 
 func transformTaskForDB(t taskModel.Task, u username.Username) taskDB {
+	if t.ID() == -1 {
+		return taskDB{0, u.Name(), t.Date(), t.Title()}
+	}
+
 	return taskDB{uint(t.ID()), u.Name(), t.Date(), t.Title()}
 }
 
@@ -63,5 +73,12 @@ func (r *TaskRepository) GetAll(u username.Username) ([]taskModel.Task, error) {
 }
 
 func (r *TaskRepository) Remove(u username.Username, id int) error {
+	if id == 0 {
+		return fmt.Errorf(IDIsNotZero)
+	}
+	if id < 0 {
+		return fmt.Errorf(InvalidID)
+	}
+
 	return r.dbHandler.Db.Delete(taskDB{ID: uint(id)}).Error
 }
